+++ src/main/scala/ulisse/infrastructures/view/common/CentralController.scala
@@ -1,4 +1,4 @@
-package ulisse.infrastructures.view.form
+package ulisse.infrastructures.view.common
 
 import ulisse.infrastructures.view.components.JComponent
